adds separate GPS events when new GPS location is detected
